pkg/authorizer: use SDK for token retrieval, exchange

+++ test/e2e_test.go
@@ -5,6 +5,7 @@ package test
 import (
 	"context"
 	"fmt"
+	"os"
 	"strings"
 	"testing"
 	"time"
@@ -53,6 +54,9 @@ func TestManagedIdentityPulls(t *testing.T) {
 		}
 
 		t.Cleanup(func() {
+			if _, skip := os.LookupEnv("SKIP_CLEANUP"); skip {
+				return
+			}
 			if err := client.Delete(ctx, &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}); err != nil {
 				t.Logf("failed to delete namespace %s: %v", namespace, err)
 			}
@@ -76,26 +80,7 @@ func TestManagedIdentityPulls(t *testing.T) {
 			t.Fatalf("failed to create Pod %s: %v", namespace, err)
 		}
 
-		EventuallyObject(t, ctx, fmt.Sprintf("Pod %s/%s to fail to pull the image", namespace, pod),
-			func(ctx context.Context) (*corev1.Pod, error) {
-				thisPod := corev1.Pod{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: pod}}
-				err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisPod), &thisPod)
-				return &thisPod, err
-			},
-			[]Predicate[*corev1.Pod]{
-				PodPhasePredicate(corev1.PodPending),
-				ConditionPredicate[*corev1.Pod](Condition{
-					Type:   string(corev1.PodReady),
-					Status: metav1.ConditionFalse,
-					Reason: "ContainersNotReady",
-				}),
-				ContainerStatePredicate("main", ContainerStateMatcher{
-					State:  ContainerStateWaiting,
-					Reason: "ImagePullBackOff",
-				}),
-			},
-			WithTimeout(2*time.Minute),
-		)
+		eventuallyFailToPullImage(t, ctx, client, namespace, pod)
 	})
 
 	t.Run("pulls succeed with acrpullbinding", func(t *testing.T) {
@@ -108,6 +93,9 @@ func TestManagedIdentityPulls(t *testing.T) {
 		}
 
 		t.Cleanup(func() {
+			if _, skip := os.LookupEnv("SKIP_CLEANUP"); skip {
+				return
+			}
 			if err := client.Delete(ctx, &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}); err != nil {
 				t.Logf("failed to delete namespace %s: %v", namespace, err)
 			}
@@ -131,20 +119,7 @@ func TestManagedIdentityPulls(t *testing.T) {
 		}); err != nil {
 			t.Fatalf("failed to create pull binding %s/%s: %v", namespace, pullBinding, err)
 		}
-		EventuallyObject(t, ctx, fmt.Sprintf("ACRPullBinding %s/%s to have credentials propagated", namespace, pullBinding),
-			func(ctx context.Context) (*msiacrpullv1beta1.AcrPullBinding, error) {
-				thisBinding := msiacrpullv1beta1.AcrPullBinding{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: pullBinding}}
-				err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisBinding), &thisBinding)
-				return &thisBinding, err
-			},
-			[]Predicate[*msiacrpullv1beta1.AcrPullBinding]{
-				func(binding *msiacrpullv1beta1.AcrPullBinding) (done bool, reasons string, err error) {
-					done = binding.Status.Error == "" && binding.Status.LastTokenRefreshTime != nil && binding.Status.TokenExpirationTime != nil
-					return done, fmt.Sprintf("wanted refresh times to be published without error, got error=%s, refresh=%s, expiration=%s", binding.Status.Error, binding.Status.LastTokenRefreshTime, binding.Status.TokenExpirationTime), nil
-				},
-			},
-			WithTimeout(2*time.Minute),
-		)
+		eventuallyFulfillPullBinding(t, ctx, client, namespace, pullBinding)
 
 		for name, image := range map[string]string{
 			"alice": cfg.AliceImage,
@@ -170,21 +145,7 @@ func TestManagedIdentityPulls(t *testing.T) {
 					t.Fatalf("failed to create Pod %s/%s: %v", namespace, name, err)
 				}
 
-				EventuallyObject(t, ctx, fmt.Sprintf("Pod %s/%s to pull the image", namespace, name),
-					func(ctx context.Context) (*corev1.Pod, error) {
-						thisPod := corev1.Pod{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: name}}
-						err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisPod), &thisPod)
-						return &thisPod, err
-					},
-					[]Predicate[*corev1.Pod]{
-						PodPhasePredicate(corev1.PodRunning),
-						ConditionPredicate[*corev1.Pod](Condition{
-							Type:   string(corev1.PodReady),
-							Status: metav1.ConditionTrue,
-						}),
-					},
-					WithTimeout(2*time.Minute),
-				)
+				eventuallyPullImage(t, ctx, client, namespace, name)
 			})
 		}
 
@@ -206,26 +167,7 @@ func TestManagedIdentityPulls(t *testing.T) {
 			t.Fatalf("failed to create Pod %s: %v", namespace, err)
 		}
 
-		EventuallyObject(t, ctx, fmt.Sprintf("Pod %s/%s to fail to pull the image", namespace, pod),
-			func(ctx context.Context) (*corev1.Pod, error) {
-				thisPod := corev1.Pod{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: pod}}
-				err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisPod), &thisPod)
-				return &thisPod, err
-			},
-			[]Predicate[*corev1.Pod]{
-				PodPhasePredicate(corev1.PodPending),
-				ConditionPredicate[*corev1.Pod](Condition{
-					Type:   string(corev1.PodReady),
-					Status: metav1.ConditionFalse,
-					Reason: "ContainersNotReady",
-				}),
-				ContainerStatePredicate("main", ContainerStateMatcher{
-					State:  ContainerStateWaiting,
-					Reason: "ImagePullBackOff",
-				}),
-			},
-			WithTimeout(2*time.Minute),
-		)
+		eventuallyFailToPullImage(t, ctx, client, namespace, pod)
 	})
 
 	t.Run("removal of acrpullbinding cleans up credentials", func(t *testing.T) {
@@ -238,6 +180,9 @@ func TestManagedIdentityPulls(t *testing.T) {
 		}
 
 		t.Cleanup(func() {
+			if _, skip := os.LookupEnv("SKIP_CLEANUP"); skip {
+				return
+			}
 			if err := client.Delete(ctx, &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}); err != nil {
 				t.Logf("failed to delete namespace %s: %v", namespace, err)
 			}
@@ -261,20 +206,7 @@ func TestManagedIdentityPulls(t *testing.T) {
 		}); err != nil {
 			t.Fatalf("failed to create pull binding %s/%s: %v", namespace, pullBinding, err)
 		}
-		EventuallyObject(t, ctx, fmt.Sprintf("ACRPullBinding %s/%s to have credentials propagated", namespace, pullBinding),
-			func(ctx context.Context) (*msiacrpullv1beta1.AcrPullBinding, error) {
-				thisBinding := msiacrpullv1beta1.AcrPullBinding{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: pullBinding}}
-				err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisBinding), &thisBinding)
-				return &thisBinding, err
-			},
-			[]Predicate[*msiacrpullv1beta1.AcrPullBinding]{
-				func(binding *msiacrpullv1beta1.AcrPullBinding) (done bool, reasons string, err error) {
-					done = binding.Status.Error == "" && binding.Status.LastTokenRefreshTime != nil && binding.Status.TokenExpirationTime != nil
-					return done, fmt.Sprintf("wanted refresh times to be published without error, got error=%s, refresh=%s, expiration=%s", binding.Status.Error, binding.Status.LastTokenRefreshTime, binding.Status.TokenExpirationTime), nil
-				},
-			},
-			WithTimeout(2*time.Minute),
-		)
+		eventuallyFulfillPullBinding(t, ctx, client, namespace, pullBinding)
 
 		const name = "alice"
 		t.Logf("creating pod %s/%s", namespace, name)
@@ -295,21 +227,7 @@ func TestManagedIdentityPulls(t *testing.T) {
 			t.Fatalf("failed to create Pod %s/%s: %v", namespace, name, err)
 		}
 
-		EventuallyObject(t, ctx, fmt.Sprintf("Pod %s/%s to pull the image", namespace, name),
-			func(ctx context.Context) (*corev1.Pod, error) {
-				thisPod := corev1.Pod{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: name}}
-				err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisPod), &thisPod)
-				return &thisPod, err
-			},
-			[]Predicate[*corev1.Pod]{
-				PodPhasePredicate(corev1.PodRunning),
-				ConditionPredicate[*corev1.Pod](Condition{
-					Type:   string(corev1.PodReady),
-					Status: metav1.ConditionTrue,
-				}),
-			},
-			WithTimeout(2*time.Minute),
-		)
+		eventuallyPullImage(t, ctx, client, namespace, name)
 
 		if err := client.Delete(ctx, &msiacrpullv1beta1.AcrPullBinding{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: pullBinding}}); err != nil {
 			t.Fatalf("failed to remove pull binding %s/%s: %v", namespace, pullBinding, err)
@@ -352,27 +270,172 @@ func TestManagedIdentityPulls(t *testing.T) {
 			t.Fatalf("failed to create Pod %s: %v", namespace, err)
 		}
 
-		EventuallyObject(t, ctx, fmt.Sprintf("Pod %s/%s to fail to pull the image", namespace, pod),
-			func(ctx context.Context) (*corev1.Pod, error) {
-				thisPod := corev1.Pod{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: pod}}
-				err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisPod), &thisPod)
-				return &thisPod, err
-			},
-			[]Predicate[*corev1.Pod]{
-				PodPhasePredicate(corev1.PodPending),
-				ConditionPredicate[*corev1.Pod](Condition{
-					Type:   string(corev1.PodReady),
-					Status: metav1.ConditionFalse,
-					Reason: "ContainersNotReady",
-				}),
-				ContainerStatePredicate("main", ContainerStateMatcher{
-					State:  ContainerStateWaiting,
-					Reason: "ImagePullBackOff",
-				}),
-			},
-			WithTimeout(2*time.Minute),
-		)
+		eventuallyFailToPullImage(t, ctx, client, namespace, pod)
 	})
+
+	t.Run("scoped acrpullbinding only allows pulls within scope", func(t *testing.T) {
+		t.Parallel()
+
+		const namespace = "scoped"
+		t.Logf("creating namespace %s", namespace)
+		if err := client.Create(ctx, &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}); err != nil && !errors.IsAlreadyExists(err) {
+			t.Fatalf("failed to create namespace %s: %v", namespace, err)
+		}
+
+		t.Cleanup(func() {
+			if _, skip := os.LookupEnv("SKIP_CLEANUP"); skip {
+				return
+			}
+			if err := client.Delete(ctx, &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}); err != nil {
+				t.Logf("failed to delete namespace %s: %v", namespace, err)
+			}
+		})
+
+		const serviceAccount = "sa"
+		t.Logf("creating service account %s/%s", namespace, serviceAccount)
+		if err := client.Create(ctx, &corev1.ServiceAccount{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: serviceAccount}}); err != nil && !errors.IsAlreadyExists(err) {
+			t.Fatalf("failed to create service account %s/%s: %v", namespace, serviceAccount, err)
+		}
+
+		const pullBinding = "pull-binding"
+		t.Logf("creating pull binding %s/%s", namespace, pullBinding)
+		if err := client.Create(ctx, &msiacrpullv1beta1.AcrPullBinding{
+			ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: pullBinding},
+			Spec: msiacrpullv1beta1.AcrPullBindingSpec{
+				AcrServer:                 cfg.RegistryFQDN,
+				ManagedIdentityResourceID: cfg.PullerResourceID,
+				ServiceAccountName:        serviceAccount,
+				Scope:                     "repository:alice:pull",
+			},
+		}); err != nil {
+			t.Fatalf("failed to create pull binding %s/%s: %v", namespace, pullBinding, err)
+		}
+		eventuallyFulfillPullBinding(t, ctx, client, namespace, pullBinding)
+
+		type imageMeta struct {
+			image   string
+			succeed bool
+		}
+		for name, imageCfg := range map[string]imageMeta{
+			"alice": {
+				image:   cfg.AliceImage,
+				succeed: true,
+			},
+			"bob": {
+				image:   cfg.BobImage,
+				succeed: false,
+			},
+		} {
+			what := "fails"
+			if imageCfg.succeed {
+				what = "succeeds"
+			}
+			t.Run(name+" "+what+" to pull the image", func(t *testing.T) {
+				t.Parallel()
+				t.Logf("creating pod %s/%s", namespace, name)
+				if err := client.Create(ctx, &corev1.Pod{
+					ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: name},
+					Spec: corev1.PodSpec{
+						Containers: []corev1.Container{{
+							Name:            "main",
+							Image:           imageCfg.image,
+							Command:         []string{"/usr/bin/sleep"},
+							Args:            []string{"infinity"},
+							ImagePullPolicy: corev1.PullAlways,
+						}},
+						ServiceAccountName: serviceAccount,
+						NodeSelector:       nodeSelector,
+					},
+				}); err != nil {
+					t.Fatalf("failed to create Pod %s/%s: %v", namespace, name, err)
+				}
+
+				if imageCfg.succeed {
+					eventuallyPullImage(t, ctx, client, namespace, name)
+				} else {
+					eventuallyFailToPullImage(t, ctx, client, namespace, name)
+				}
+			})
+		}
+
+		const pod = "fail"
+		t.Logf("creating pod without service account %s/%s", namespace, pod)
+		if err := client.Create(ctx, &corev1.Pod{
+			ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: pod},
+			Spec: corev1.PodSpec{
+				Containers: []corev1.Container{{
+					Name:            "main",
+					Image:           cfg.AliceImage,
+					Command:         []string{"/usr/bin/sleep"},
+					Args:            []string{"infinity"},
+					ImagePullPolicy: corev1.PullAlways,
+				}},
+				NodeSelector: nodeSelector,
+			},
+		}); err != nil {
+			t.Fatalf("failed to create Pod %s: %v", namespace, err)
+		}
+
+		eventuallyFailToPullImage(t, ctx, client, namespace, pod)
+	})
+}
+
+func eventuallyFulfillPullBinding(t *testing.T, ctx context.Context, client crclient.Client, namespace, name string) {
+	EventuallyObject(t, ctx, fmt.Sprintf("ACRPullBinding %s/%s to have credentials propagated", namespace, name),
+		func(ctx context.Context) (*msiacrpullv1beta1.AcrPullBinding, error) {
+			thisBinding := msiacrpullv1beta1.AcrPullBinding{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: name}}
+			err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisBinding), &thisBinding)
+			return &thisBinding, err
+		},
+		[]Predicate[*msiacrpullv1beta1.AcrPullBinding]{
+			func(binding *msiacrpullv1beta1.AcrPullBinding) (done bool, reasons string, err error) {
+				done = binding.Status.Error == "" && binding.Status.LastTokenRefreshTime != nil && binding.Status.TokenExpirationTime != nil
+				return done, fmt.Sprintf("wanted refresh times to be published without error, got error=%s, refresh=%s, expiration=%s", binding.Status.Error, binding.Status.LastTokenRefreshTime, binding.Status.TokenExpirationTime), nil
+			},
+		},
+		WithTimeout(2*time.Minute),
+	)
+}
+
+func eventuallyPullImage(t *testing.T, ctx context.Context, client crclient.Client, namespace, name string) {
+	EventuallyObject(t, ctx, fmt.Sprintf("Pod %s/%s to pull the image", namespace, name),
+		func(ctx context.Context) (*corev1.Pod, error) {
+			thisPod := corev1.Pod{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: name}}
+			err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisPod), &thisPod)
+			return &thisPod, err
+		},
+		[]Predicate[*corev1.Pod]{
+			PodPhasePredicate(corev1.PodRunning),
+			ConditionPredicate[*corev1.Pod](Condition{
+				Type:   string(corev1.PodReady),
+				Status: metav1.ConditionTrue,
+			}),
+		},
+		WithTimeout(2*time.Minute),
+	)
+}
+
+func eventuallyFailToPullImage(t *testing.T, ctx context.Context, client crclient.Client, namespace, name string) {
+	EventuallyObject(t, ctx, fmt.Sprintf("Pod %s/%s to fail to pull the image", namespace, name),
+		func(ctx context.Context) (*corev1.Pod, error) {
+			thisPod := corev1.Pod{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: name}}
+			err := client.Get(ctx, crclient.ObjectKeyFromObject(&thisPod), &thisPod)
+			return &thisPod, err
+		},
+		[]Predicate[*corev1.Pod]{
+			PodPhasePredicate(corev1.PodPending),
+			ConditionPredicate[*corev1.Pod](Condition{
+				Type:   string(corev1.PodReady),
+				Status: metav1.ConditionFalse,
+				Reason: "ContainersNotReady",
+			}),
+			ContainerStatePredicate("main", ContainerStateMatcher{
+				State:  ContainerStateWaiting,
+				Reason: "ImagePullBackOff",
+			}),
+		},
+		WithTimeout(2*time.Minute),
+	)
 }
 
 // PodPhasePredicate returns a predicate that validates that a pod is in the requisite phase.
